refactor(structure): remove legacy folders and unify modular architecture
