--- backend/src/routes/auth.routes.js
@@ -1,9 +1,0 @@
-import { Router } from "express";
-import { register, login } from "../controllers/auth.controller.js";
-
-const router = Router();
-
-router.post("/register", register);
-router.post("/login", login);
-
-export default router;
